fixed exception problem

+++ backend/src/main/java/com/example/demo/domain/blogpost/BlogPostServiceImpl.java
@@ -15,6 +15,7 @@ import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 
+import javax.transaction.Transactional;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
@@ -31,6 +32,7 @@ public class BlogPostServiceImpl implements BlogPostService {
     private UserService userService;
 
     @Override
+    @Transactional
     public BlogPost create(BlogPost newEntity) {
         log.info("Attempting to save Entry {}", newEntity);
         newEntity.setCreationTime(LocalDateTime.now());
@@ -41,6 +43,7 @@ public class BlogPostServiceImpl implements BlogPostService {
     }
 
     @Override
+    @Transactional
     public List<BlogPost> getAll() {
         log.info("Attempting to find All Entries");
         List<BlogPost> foundItems = (List<BlogPost>) repository.findAll();
@@ -49,6 +52,7 @@ public class BlogPostServiceImpl implements BlogPostService {
     }
 
     @Override
+    @Transactional
     public List<BlogPost> getAllWithLimitAfterId(UUID blogId, long limit) {
         BlogPost post = getById(blogId);
 
@@ -59,14 +63,16 @@ public class BlogPostServiceImpl implements BlogPostService {
     }
 
     @Override
+    @Transactional
     public List<BlogPost> getAllFromPageWithLimit(long page, long limit) {
         log.info("Attempting to find Entries from page {} with limit {}", page, limit);
-        List<BlogPost> foundItems = (List<BlogPost>) repository.findByLimitAndWithOffset(limit, limit * page);
+        List<BlogPost> foundItems = repository.findByLimitAndWithOffset(limit, limit * page);
         log.info("Successfully found Entries from page {} with limit {}", page, limit);
         return foundItems;
     }
 
     @Override
+    @Transactional
     public BlogPost getById(UUID blogId) {
         log.info("Attempting to find entry with id {}", blogId);
         Optional<BlogPost> post = repository.findById(blogId);
@@ -79,10 +85,14 @@ public class BlogPostServiceImpl implements BlogPostService {
     }
 
     @Override
+    @Transactional
     public BlogPost updateById(UUID blogId, BlogPost newEntity) {
         // Already Covered By Tests
         BlogPost post = getById(blogId);
-        if (isOwnPost(post)) {
+
+        if (!isOwnPost(post)) {
+            throw new AuthorizationServiceException("This Post does not belong to the User");
+        } else {
             post.setCategory(newEntity.getCategory());
             post.setText(newEntity.getText());
             post.setTitle(newEntity.getTitle());
@@ -91,16 +101,20 @@ public class BlogPostServiceImpl implements BlogPostService {
             BlogPost updatedPost = repository.save(post);
             log.info("BlogPost with id {} updated to {}", blogId, updatedPost);
             return updatedPost;
-        } else {
-            throw new AuthorizationServiceException("The Post ");
         }
     }
 
     @Override
+    @Transactional
     public void deleteById(UUID blogId) {
-        log.info("Attempting to delete entry with id {}", blogId);
-        repository.deleteById(blogId);
-        log.info("Successfully deleted entry with id {}", blogId);
+        BlogPost post = getById(blogId);
+        if (!isOwnPost(post)) {
+            throw new AuthorizationServiceException("This Post does not belong to the User");
+        } else {
+            log.info("Attempting to delete entry with id {}", blogId);
+            repository.deleteById(blogId);
+            log.info("Successfully deleted entry with id {}", blogId);
+        }
     }
 
     public boolean isOwnPost(BlogPost post) {
